+++ apps/ymir-sdl3/src/app/shared_context.cpp
@@ -4,9 +4,6 @@
 
 #include <ymir/sys/saturn.hpp>
 
-#include <ymir/util/scope_guard.hpp>
-
-#include <unordered_set>
 #include <vector>
 
 namespace app {
@@ -85,6 +82,22 @@ std::filesystem::path SharedContext::GetPerGameExternalBackupRAMPath(ymir::bup::
     const std::filesystem::path basePath = profile.GetPath(ProfilePath::BackupMemory) / "games";
     std::filesystem::create_directories(basePath);
     return basePath / fmt::format("bup-ext-{}M-{}.bin", BupSizeToSize(bupSize) * 8 / 1024 / 1024, GetGameFileName());
+}
+
+std::filesystem::path SharedContext::GetPersistentSMPCDataPath() const {
+    const char *regionSuffix;
+    if (iplRomInfo != nullptr) {
+        switch (iplRomInfo->region) {
+        case ymir::db::SystemRegion::US_EU: regionSuffix = "us_eu"; break;
+        case ymir::db::SystemRegion::JP: regionSuffix = "jp"; break;
+        case ymir::db::SystemRegion::KR: regionSuffix = "asia"; break;
+        default: regionSuffix = "other"; break;
+        }
+    } else {
+        regionSuffix = "none";
+    }
+
+    return profile.GetPath(ProfilePath::PersistentState) / fmt::format("smpc-{}.bin", regionSuffix);
 }
 
 SDL_DisplayID SharedContext::GetSelectedDisplay() const {